@@ -1,24 +1,25 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="MediatR" Version="12.1.1" />
-    <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="7.0.5" />
-	<PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
-	<PackageReference Include="OpenTelemetry" Version="1.5.1" />
-	<PackageReference Include="OpenTelemetry.Exporter.Console" Version="1.5.1" />
-	<PackageReference Include="OpenTelemetry.Exporter.OpenTelemetryProtocol" Version="1.5.1" />
+    <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="8.0.0-rc.1.23421.29" />
+	<PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
+	<PackageReference Include="OpenTelemetry" Version="1.6.0" />
+	<PackageReference Include="OpenTelemetry.Exporter.Console" Version="1.6.0" />
+	<PackageReference Include="OpenTelemetry.Exporter.OpenTelemetryProtocol" Version="1.6.0" />
 	<PackageReference Include="OpenTelemetry.Exporter.OpenTelemetryProtocol.Logs" Version="1.5.0-rc.1" />
-	<PackageReference Include="OpenTelemetry.Extensions.Hosting" Version="1.5.1" />
-	<PackageReference Include="OpenTelemetry.Instrumentation.Http" Version="1.5.0-beta.1" />
+	<PackageReference Include="OpenTelemetry.Extensions.Hosting" Version="1.6.0" />
+	<PackageReference Include="OpenTelemetry.Instrumentation.Http" Version="1.5.1-beta.1" />
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\sdk\Townsharp.Core\Townsharp.Core.csproj" />
     <ProjectReference Include="..\..\sdk\Townsharp.Infrastructure\Townsharp.Infrastructure.csproj" />
     <ProjectReference Include="..\..\sdk\Townsharp\Townsharp.csproj" />
   </ItemGroup>
