@@ -2,16 +2,16 @@
 
 namespace Townsharp.Infrastructure.GameConsoles;
 
-public class GameConsoleEvent
+public class ConsoleEvent
 {
     private readonly JsonNode result;
 
-    public GameConsoleEvent(JsonNode result)
+    public ConsoleEvent(JsonNode result)
     {
         this.result = result;
     }
 
     public JsonNode Result => this.result;
 
-    public override string ToString() => Result.ToJsonString(new System.Text.Json.JsonSerializerOptions { WriteIndented = true });
+    public override string ToString() => this.Result.ToJsonString(new System.Text.Json.JsonSerializerOptions { WriteIndented = true });
 }
